+++ app_test.go
@@ -276,28 +276,6 @@ func TestInvokes(t *testing.T) {
 	})
 }
 
-func TestSupply(t *testing.T) {
-	t.Run("NothingIsSupplied", func(t *testing.T) {
-		app := fxtest.New(t, Supply())
-		defer app.RequireStart().RequireStop()
-		require.NoError(t, app.Err())
-	})
-
-	t.Run("SomethingIsSupplied", func(t *testing.T) {
-		type A struct{}
-		type B struct{}
-
-		aIn, bIn := A{}, B{}
-		aOut, bOut := (*A)(nil), (*B)(nil)
-		app := fxtest.New(t, Supply(&aIn, &bIn), Populate(&aOut, &bOut))
-		defer app.RequireStart().RequireStop()
-
-		require.NoError(t, app.Err())
-		require.Equal(t, &aIn, aOut)
-		require.Equal(t, &bIn, bOut)
-	})
-}
-
 func TestError(t *testing.T) {
 	t.Run("NilErrorOption", func(t *testing.T) {
 		var invoked bool
